+++ src/main/java/org/mybatis/dynamic/sql/SqlTable.java
@@ -18,93 +18,19 @@ package org.mybatis.dynamic.sql;
 import java.sql.JDBCType;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.function.Supplier;
 
 import org.jetbrains.annotations.NotNull;
 
 public class SqlTable implements TableExpression {
 
-    protected Supplier<String> nameSupplier;
+    protected String tableName;
 
     protected SqlTable(String tableName) {
-        Objects.requireNonNull(tableName);
-
-        this.nameSupplier = () -> tableName;
+        this.tableName = Objects.requireNonNull(tableName);
     }
 
-    /**
-     * Creates an SqlTable whose name can be changed at runtime.
-     *
-     * @param tableNameSupplier table name supplier
-     * @deprecated please use {@link AliasableSqlTable} if you need to change the table name at runtime
-     */
-    @Deprecated
-    protected SqlTable(Supplier<String> tableNameSupplier) {
-        Objects.requireNonNull(tableNameSupplier);
-
-        this.nameSupplier = tableNameSupplier;
-    }
-
-    /**
-     * Creates an SqlTable whose name can be changed at runtime.
-     *
-     * @param schemaSupplier schema supplier
-     * @param tableName table name
-     * @deprecated please use {@link AliasableSqlTable} if you need to change the table name at runtime
-     */
-    @Deprecated
-    protected SqlTable(Supplier<Optional<String>> schemaSupplier, String tableName) {
-        this(Optional::empty, schemaSupplier, tableName);
-    }
-
-    /**
-     * Creates an SqlTable whose name can be changed at runtime.
-     *
-     * @param catalogSupplier catalog supplier
-     * @param schemaSupplier schema supplier
-     * @param tableName table name
-     * @deprecated please use {@link AliasableSqlTable} if you need to change the table name at runtime
-     */
-    @Deprecated
-    protected SqlTable(Supplier<Optional<String>> catalogSupplier, Supplier<Optional<String>> schemaSupplier,
-            String tableName) {
-        Objects.requireNonNull(catalogSupplier);
-        Objects.requireNonNull(schemaSupplier);
-        Objects.requireNonNull(tableName);
-
-        this.nameSupplier = () -> compose(catalogSupplier, schemaSupplier, tableName);
-    }
-
-    private String compose(Supplier<Optional<String>> catalogSupplier, Supplier<Optional<String>> schemaSupplier,
-            String tableName) {
-        return catalogSupplier.get().map(c -> compose(c, schemaSupplier, tableName))
-                .orElseGet(() -> compose(schemaSupplier, tableName));
-    }
-
-    private String compose(String catalog, Supplier<Optional<String>> schemaSupplier, String tableName) {
-        return schemaSupplier.get().map(s -> composeCatalogSchemaAndTable(catalog, s, tableName))
-                .orElseGet(() -> composeCatalogAndTable(catalog, tableName));
-    }
-
-    private String compose(Supplier<Optional<String>> schemaSupplier, String tableName) {
-        return schemaSupplier.get().map(s -> composeSchemaAndTable(s, tableName))
-                .orElse(tableName);
-    }
-
-    private String composeCatalogAndTable(String catalog, String tableName) {
-        return catalog + ".." + tableName; //$NON-NLS-1$
-    }
-
-    private String composeSchemaAndTable(String schema, String tableName) {
-        return schema + "." + tableName; //$NON-NLS-1$
-    }
-
-    private String composeCatalogSchemaAndTable(String catalog, String schema, String tableName) {
-        return catalog + "." + schema + "." + tableName; //$NON-NLS-1$ //$NON-NLS-2$
-    }
-
-    public String tableNameAtRuntime() {
-        return nameSupplier.get();
+    public String tableName() {
+        return tableName;
     }
 
     public BasicColumn allColumns() {